item process permission

+++ routes/web.php
@@ -102,18 +102,18 @@ Route::middleware(['auth:company_user'])->prefix('user')->group(function () {
     });
 
     Route::prefix('process-item/{item}')->group(function () {
-        Route::get('/', [ProcessController::class, 'index'])->name('process.index');
+        Route::get('/', [ProcessController::class, 'index'])->name('process.index')->middleware('permission:item_process,view');
         // Route::get('/create', [ProcessController::class, 'create'])->name('process.create');
-        Route::get('/edit/{id}', [ProcessController::class, 'edit'])->name('process.edit');
-        Route::put('/update/{id}', [ProcessController::class, 'update'])->name('process.update');
-        Route::delete('/delete/{id}', [ProcessController::class, 'destroy'])->name('process.delete');
+        Route::get('/edit/{id}', [ProcessController::class, 'edit'])->name('process.edit')->middleware('permission:item_process,edit');
+        Route::put('/update/{id}', [ProcessController::class, 'update'])->name('process.update')->middleware('permission:item_process,edit');
+        Route::delete('/delete/{id}', [ProcessController::class, 'destroy'])->name('process.delete')->middleware('permission:item_process,delete');
 
         // Drag & Drop Sort
         Route::post('/sort', [ProcessController::class, 'sort'])->name('process.sort');
     });
-    Route::post('process-item/store-multiple/{itemId}', [ProcessController::class, 'storeMultiple'])->name('process.storeMultiple');
+    Route::post('process-item/store-multiple/{itemId}', [ProcessController::class, 'storeMultiple'])->name('process.storeMultiple')->middleware('permission:item_process,add');
 
-    Route::get('/process-item', [ProcessController::class, 'itemList'])->name('process.items');
+    Route::get('/process-item', [ProcessController::class, 'itemList'])->name('process.items')->middleware('permission:item_process,view');
     Route::get('/process/get/{id}', [ProcessController::class, 'getProcesses'])->name('process.get');
 
     Route::prefix('process_master')->group(function () {
